Make IntegrationSettings properly immutable and remove ImmutableIntegrationSettings
